added some child read stuff to the new trigger logic
git-svn-id: svn://orca.physics.unc.edu/Orca@4074 fd86f90f-2c0a-422a-b972-84106d8db31b

+++ Source/Objects/TriggerSystem/Logic/LogicInBit/ORLogicInBitModel.h
@@ -21,7 +21,7 @@
 
 @interface ORLogicInBitModel :  OrcaObject <TriggerBitReading>
 {
-	unsigned long bit;
+	unsigned short bit;
 }
 - (unsigned short) bit;
 - (void) setBit:(unsigned short)aBit;
